// src/modules/cases/case-items/case-items.types.ts
import { Document, Types } from "mongoose";

/**
 * Интерфейс для модели CaseItem в базе данных
 *
 * CaseItem представляет связь между кейсом (Case) и предметом (Item)
 * Это связующая таблица (join table) для реализации связи "многие-ко-многим"
 *
 * Один кейс может содержать множество предметов (через разные записи CaseItem)
 * Один предмет может быть в разных кейсах (через разные записи CaseItem)
 */
export interface ICaseItem extends Document {
  // ID кейса, в котором может выпасть предмет
  caseId: Types.ObjectId;

  // ID предмета, который может выпасть из кейса
  itemId: Types.ObjectId;

  // Шанс выпадения этого предмета из конкретного кейса (в процентах)
  // Например, chance: 50 означает 50% вероятность выпадения
  chance: number; // %
}